+++ lib/ui/pages/splash_page.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:airplane_apps/shared/theme.dart';
+
+class SplashPage extends StatelessWidget {
+  const SplashPage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: kPrimaryColor,
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Container(
+              width: 100,
+              height: 100,
+              margin: EdgeInsets.only(bottom: 50),
+              decoration: BoxDecoration(
+                image: DecorationImage(
+                  image: AssetImage('assets/images/icon_plane.png'), 
+                ),
+              ),
+            ),
+            Text(
+              'AIRPLANE',
+              style: whiteTextStyle.copyWith(
+                fontSize: 32,
+                fontWeight: medium,
+                letterSpacing: 10,
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
